requirements.txt
requests
slack_sdk
googletrans==4.0.0-rc1
PyYAML